Omni-Optimizer

Simple interface to the C-implementation of the Omni-optimizer by Deb and Tiwari [1,2]. The algorithm “is designed as a generic multi-objective, multi-optima optimizer” [2].

omniopt(
  fn,
  pop.size = 4L,
  n.gens = 100L,
  p.cross = 0.6,
  p.mut = 1/smoof::getNumberOfParameters(fn),
  eta.cross = 20L,
  eta.mut = 20L,
  mate = "normal",
  delta = 0.001,
  var.space.niching = FALSE,
  obj.space.niching = TRUE,
  init = "random",
  frequency = 1,
  seed = runif(1),
  envir = environment()
)

Arguments

fn
[function]
Multi-objective function of type smoof_function (see smoof).

pop.size
[integer(1)]
Population size. Must be a multiple of 4. The default is 4.

n.gens
[integer(1)]
The number of generations (stopping condition). Defaults to 100.

p.cross
[numeric(1)]
Probability of crossover (within \([0.6, 1.0]\)). Defaults to 0.6.

p.mut
[numeric(1)]
Probablity of mutation (within \([0,1]\)). Default to \(1/n\) where \(n\) is the number of decision variables of fn.

eta.cross
[integer(1)]
Value of distribution index for crossover in \(\{5, \ldots, 20\}\). Default is 20.

eta.mut
[integer(1)]
Value of distribution index for mutation in \(\{5, \ldots, 50\}\). Default is 20.

mate
[character(1)]
Choice for selection restriction. Either “normal” for normal selection or “restricted” for restricted selection.

delta
[numeric(1)]
Value \(\delta \in (0.0,1.0)\) for loose domination. Default is \(0.001\).

var.space.niching
[logical(1)]
Use variable space niching? Default is FALSE. Note that at least on of var.space.niching or obj.space.niching must be TRUE.

obj.space.niching
[logical(1)]
Use objective space niching? Default is TRUE. Note that at least on of var.space.niching or obj.space.niching must be TRUE.

init
[character(1)]
This parameter determines how to initialize the population: “random” (the default) for uniform random generation and “lhs” for Latin-Hypercube-Sampling (LHS).

frequency
[integer(1)]
Frequency with which the population information is to be stored. Defaults to 1.

seed
[numeric(1)]
Single numeric value in \([0,1]\). Defaults to a random number within this interval.

envir
[environment]
This parameter is required for calling R functions from C. Do not change the default unless you know what you are doing!

Value

[List] List with the following entries:

pareto.set

Matrix of non-dominated points (each column is a point).

pareto.front

Matrix of non-dominated point objective values (each column is a point).

Details

The function expects a real-valued benchmark function from package smoof, the population size and the number of generations (the only stopping condition) as mandatory arguments. Besides there are various parameters that can be adjusted (see the referenced papers for an in-depth explanation of the algorithms working principles).

The original C-code can be found at the COIN laboratory website.

References

[1] Kalyanmoy Deb, Santosh Tiwari: Omni-optimizer: A generic evolutionary algorithm for single and multi-objective optimization. European Journal of Operations Research 185(3): 1062-1087.

[2] Kalyanmoy Deb, Santosh Tiwari: Omni-optimizer: A Procedure for Single and Multi-objective Optimization. In: Proceedings of the Evolutionary Multi-Criterion Conference (EMO) 2005: 47-61.

Examples

if (FALSE) {
library(smoof)

fn = smoof::makeDTLZ1Function(dimensions = 2L, n.objectives = 2L)
fn = smoof::addLoggingWrapper(fn, logg.x = TRUE, logg.y = TRUE)

set.seed(52357) # reproducibility
res = omniopt(fn, pop.size = 20L, n.gens = 1000L)
print(res)
print(smoof::getLoggedValues(fn, compact = TRUE))
plot(t(res$pareto.front))
}